+++ packages/drippie-mon/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @eth-optimism/drippie-mon
 
+## 0.3.21
+
+### Patch Changes
+
+- Updated dependencies [1e76cdb86]
+  - @eth-optimism/core-utils@0.11.0
+  - @eth-optimism/common-ts@0.6.7
+  - @eth-optimism/sdk@1.6.10
+
+## 0.3.20
+
+### Patch Changes
+
+- @eth-optimism/sdk@1.6.9
+
 ## 0.3.19
 
 ### Patch Changes